pyproject.toml
[project]
name = "bao-ai"
dynamic = ["version"]
description = "A personal AI assistant that remembers, learns from experience, and never makes the same mistake twice."
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Bao contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.20.0,<1.0.0",
    "openai>=1.83.0,<3.0.0",
    "anthropic>=0.39.0,<1.0.0",
    "google-genai>=1.0.0,<2.0.0",
    "pydantic>=2.12.0,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "websockets>=13.0,<17.0",
    "websocket-client>=1.9.0,<2.0.0",
    "httpx>=0.28.0,<1.0.0",
    "oauth-cli-kit>=0.1.3,<1.0.0",
    "loguru>=0.7.3,<1.0.0",
    "readability-lxml>=0.8.4,<1.0.0",
    "rich>=14.0.0,<15.0.0",
    "croniter>=6.0.0,<7.0.0",
    "dingtalk-stream>=0.24.0,<1.0.0",
    "python-telegram-bot[socks]>=22.6,<23.0",
    "lark-oapi>=1.5.0,<2.0.0",
    "socksio>=1.0.0,<2.0.0",
    "python-socketio>=5.16.0,<6.0.0",
    "msgpack>=1.1.0,<2.0.0",
    "slack-sdk>=3.39.0,<4.0.0",
    "slackify-markdown>=0.2.0,<1.0.0",
    "qq-botpy>=1.2.0,<2.0.0",
    "python-socks[asyncio]>=2.8.0,<3.0.0",
    "prompt-toolkit>=3.0.50,<4.0.0",
    "mcp>=1.26.0,<2.0.0",
    "json-repair>=0.57.0,<1.0.0",
    "lancedb>=0.20.0,<0.26.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.18.0",
]

[project.optional-dependencies]
desktop = [
    "PySide6-Essentials>=6.7.0,<7.0.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
]
desktop-automation = [
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
]
desktop-build = [
    "PySide6-Essentials>=6.7.0,<7.0.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "nuitka>=4.0.2,<5.0.0",
    "ordered-set>=4.1.0",
    "zstandard>=0.25.0",
]
desktop-build-pyinstaller = [
    "PySide6-Essentials>=6.7.0,<7.0.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "pyinstaller>=6.19.0,<7.0.0",
]
dev = [
    "pytest>=9.0.0,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
bao = "bao.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bao"]

[tool.hatch.build.targets.wheel.sources]
"bao" = "bao"

[tool.hatch.version]
path = "bao/__about__.py"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "bao/**/*.py",
    "bao/skills/**/*.md",
    "bao/skills/**/*.sh",
    "bao/templates/**/*.md",
]

[tool.hatch.build.targets.sdist]
include = [
    "bao/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "bao/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["N802", "N815"]
"app/backend/*.py" = ["N802", "N815"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
strict_markers = true
markers = [
    "unit: fast deterministic tests for local iteration",
    "integration: cross-module and boundary tests",
    "smoke: critical-path regression checks",
    "desktop_ui_smoke: desktop UI smoke and screenshot contract checks",
    "gui: desktop and QML tests that need PySide6 or UI harnesses",
    "slow: expensive tests that are not part of the default fast path",
]
